refactor(agglayer): share GER note construction between UPDATE_GER and REMOVE_GER
UpdateGerNote::create and RemoveGerNote::create built structurally identical notes
(8 felts of GER storage, a network-account target on the bridge, public metadata, no
assets), differing only in the note script. Extract the shared body into a crate-internal
create_ger_note helper that each note type calls with its own script.
