+++ crates/miden-agglayer/src/lib.rs
@@ -25,6 +25,7 @@ pub mod config_note;
 pub mod errors;
 pub mod eth_types;
 pub mod faucet;
+mod ger_note;
 pub mod remove_ger_note;
 #[cfg(feature = "testing")]
 pub mod testing;
